flow control improvement

+++ src_utils/flow_control.c
@@ -6,7 +6,7 @@
 /*   By: omoreno- <omoreno-@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/13 19:06:22 by omoreno-          #+#    #+#             */
-/*   Updated: 2022/12/13 19:35:27 by omoreno-         ###   ########.fr       */
+/*   Updated: 2022/12/14 10:58:32 by omoreno-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 static t_feedback	g_feedback;
 
-int	ft_get_feedback_pending(void)
+size_t	ft_get_feedback_pending(void)
 {
 	return (g_feedback.pending);
 }
@@ -24,7 +24,7 @@ int	ft_get_queue_hi(void)
 	return (g_feedback.queue_hi);
 }
 
-void	ft_set_feedback_pending(int val)
+void	ft_set_feedback_pending(size_t val)
 {
 	g_feedback.pending = val;
 }
